Share the POSIX libc bindings across the UNIX FFM classes
Five of the six *LibcFunctions classes carried byte-identical bindings
for getpid, getrlimit with its struct layout and both accessors, and
gethostname. Verified identical rather than merely similarly named: the
AIX and Solaris blocks diff clean against each other, and OpenBSD
differs only by not binding gethostname.

They now live in PosixLibcFunctions, which the five extend. Java statics
are inherited by name, so all 22 call sites compile unchanged.

The class is named for its membership rule rather than its members.
Handles resolve with findOrThrow in a static initializer, so a symbol
missing on one platform raises ExceptionInInitializerError there and
takes the rest of the class with it. POSIX-mandated is what guarantees
the symbol exists everywhere, which makes it the criterion for belonging
rather than a matter of taste.

getloadavg is the counterexample and stays per-platform. It looks like a
peer of these and four of the six bind it, but it is a BSD extension
rather than POSIX and AIX does not have it -- AIX reads load average
from perfstat_cpu_total. Hoisting it would have broken AIX at class
load, on a platform that never asks for it. A BsdLibcFunctions tier to
hold it was considered and rejected: it would hold exactly one member,
and the shareable set does not otherwise cluster by family, since AIX
and Solaris are identical to the BSDs for everything in the POSIX tier.

RLIMIT_NOFILE stays in each subclass: POSIX names it but does not fix
its value, which is 7 on Linux and AIX, 8 on the BSDs and 5 on
Solaris/illumos. Its per-platform javadoc already records that JNA's
inherited Linux value would query the wrong resource.

DragonFly BSD is untouched. It binds only lwp_gettid, so it has nothing
to inherit and extending would bind three handles it never calls.

Linux needed one extra step: it assigns its handles in a static block
rather than per-field initializers, so the three assignments came out
along with the declarations.

The JNA side needs no equivalent. jna-platform's LibCAPI already
declares gethostname, getrlimit and getloadavg, and OSHI's own CLibrary
declares getpid, so the platform interfaces inherit them; the only thing
duplicated there is the RLIMIT_NOFILE constant, which must be. The
asymmetry is that JNA inherits its bindings while FFM hand-writes them.
