+++ oshi-core-ffm/src/main/java/oshi/ffm/platform/unix/aix/AixLibcFunctions.java
@@ -9,16 +9,12 @@ import static java.lang.foreign.ValueLayout.JAVA_INT;
 import static java.lang.foreign.ValueLayout.JAVA_LONG;
 
 import java.lang.foreign.FunctionDescriptor;
-import java.lang.foreign.MemoryLayout;
-import java.lang.foreign.MemoryLayout.PathElement;
 import java.lang.foreign.MemorySegment;
-import java.lang.foreign.StructLayout;
 import java.lang.foreign.SymbolLookup;
 import java.lang.foreign.ValueLayout;
 import java.lang.invoke.MethodHandle;
-import java.lang.invoke.VarHandle;
 
-import oshi.ffm.ForeignFunctions;
+import oshi.ffm.platform.unix.PosixLibcFunctions;
 
 /**
  * FFM bindings for AIX libc functions used by OSHI.
@@ -26,7 +22,7 @@ import oshi.ffm.ForeignFunctions;
  * Most performance and configuration data on AIX comes from {@code libperfstat} (see {@link PerfstatFunctions}); the
  * libc surface here is limited to thread/process IDs, hostname, and {@code getrlimit}.
  */
-public final class AixLibcFunctions extends ForeignFunctions {
+public final class AixLibcFunctions extends PosixLibcFunctions {
 
     private AixLibcFunctions() {
     }
@@ -37,29 +33,8 @@ public final class AixLibcFunctions extends ForeignFunctions {
     /** {@code getrlimit} resource: maximum number of open file descriptors. AIX value (7). */
     public static final int RLIMIT_NOFILE = 7;
 
-    /** Layout of AIX {@code struct rlimit}: two {@code rlim_t} (LP64 long) fields. */
-    public static final StructLayout RLIMIT_LAYOUT = MemoryLayout.structLayout(JAVA_LONG.withName("rlim_cur"),
-            JAVA_LONG.withName("rlim_max"));
-
-    private static final VarHandle RLIMIT_CUR = RLIMIT_LAYOUT.varHandle(PathElement.groupElement("rlim_cur"));
-    private static final VarHandle RLIMIT_MAX = RLIMIT_LAYOUT.varHandle(PathElement.groupElement("rlim_max"));
-
     // libc is already loaded into the JVM process; defaultLookup() avoids versioning pitfalls.
     private static final SymbolLookup LIBC = LINKER.defaultLookup();
-
-    // pid_t getpid(void);
-    private static final MethodHandle getpid = LINKER.downcallHandle(LIBC.findOrThrow("getpid"),
-            FunctionDescriptor.of(JAVA_INT));
-
-    /**
-     * Calls {@code getpid()}.
-     *
-     * @return the process ID of the calling process
-     * @throws Throwable on FFM invocation error
-     */
-    public static int getpid() throws Throwable {
-        return (int) getpid.invokeExact();
-    }
 
     // tid_t thread_self(void); // AIX-specific
     private static final MethodHandle thread_self = LINKER.downcallHandle(LIBC.findOrThrow("thread_self"),
@@ -73,42 +48,6 @@ public final class AixLibcFunctions extends ForeignFunctions {
      */
     public static int thread_self() throws Throwable {
         return (int) thread_self.invokeExact();
-    }
-
-    // int getrlimit(int resource, struct rlimit *rlim);
-    private static final MethodHandle getrlimit = LINKER.downcallHandle(LIBC.findOrThrow("getrlimit"),
-            FunctionDescriptor.of(JAVA_INT, JAVA_INT, ADDRESS));
-
-    /**
-     * Calls {@code getrlimit(resource, rlim)}.
-     *
-     * @param resource resource constant (e.g. {@link #RLIMIT_NOFILE})
-     * @param rlim     segment allocated with {@link #RLIMIT_LAYOUT}
-     * @return 0 on success, -1 on error
-     * @throws Throwable on FFM invocation error
-     */
-    public static int getrlimit(int resource, MemorySegment rlim) throws Throwable {
-        return (int) getrlimit.invokeExact(resource, rlim);
-    }
-
-    /**
-     * Reads {@code rlim_cur} from an rlimit segment populated by {@link #getrlimit(int, MemorySegment)}.
-     *
-     * @param rlim segment allocated with {@link #RLIMIT_LAYOUT}
-     * @return the soft resource limit
-     */
-    public static long rlimitCur(MemorySegment rlim) {
-        return (long) RLIMIT_CUR.get(rlim, 0L);
-    }
-
-    /**
-     * Reads {@code rlim_max} from an rlimit segment populated by {@link #getrlimit(int, MemorySegment)}.
-     *
-     * @param rlim segment allocated with {@link #RLIMIT_LAYOUT}
-     * @return the hard resource limit
-     */
-    public static long rlimitMax(MemorySegment rlim) {
-        return (long) RLIMIT_MAX.get(rlim, 0L);
     }
 
     // int open(const char *path, int flags);
@@ -160,19 +99,4 @@ public final class AixLibcFunctions extends ForeignFunctions {
         return (long) pread.invokeExact(fd, buf, count, offset);
     }
 
-    // int gethostname(char *name, size_t namelen);
-    private static final MethodHandle gethostname = LINKER.downcallHandle(LIBC.findOrThrow("gethostname"),
-            FunctionDescriptor.of(JAVA_INT, ADDRESS, SIZE_T));
-
-    /**
-     * Calls {@code gethostname(name, namelen)}.
-     *
-     * @param name    buffer for the hostname (allocated by caller)
-     * @param namelen size of {@code name} in bytes
-     * @return 0 on success, -1 on error
-     * @throws Throwable on FFM invocation error
-     */
-    public static int gethostname(MemorySegment name, long namelen) throws Throwable {
-        return (int) gethostname.invokeExact(name, namelen);
-    }
 }
