+++ oshi-core-ffm/src/main/java/oshi/ffm/platform/unix/solaris/SolarisLibcFunctions.java
@@ -20,7 +20,7 @@ import java.lang.foreign.ValueLayout;
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.VarHandle;
 
-import oshi.ffm.ForeignFunctions;
+import oshi.ffm.platform.unix.PosixLibcFunctions;
 
 /**
  * FFM bindings for Solaris/illumos libc functions used by OSHI.
@@ -29,7 +29,7 @@ import oshi.ffm.ForeignFunctions;
  * and {@code getrlimit}. The {@code RLIMIT_NOFILE} resource constant is {@code 5} on Solaris/illumos (not {@code 7} as
  * on Linux — see <a href="https://illumos.org/man/2/getrlimit">illumos getrlimit(2)</a>).
  */
-public final class SolarisLibcFunctions extends ForeignFunctions {
+public final class SolarisLibcFunctions extends PosixLibcFunctions {
 
     private SolarisLibcFunctions() {
     }
@@ -45,29 +45,8 @@ public final class SolarisLibcFunctions extends ForeignFunctions {
      */
     public static final int RLIMIT_NOFILE = 5;
 
-    /** Layout of Solaris/illumos {@code struct rlimit}: two {@code rlim_t} (LP64 long) fields. */
-    public static final StructLayout RLIMIT_LAYOUT = MemoryLayout.structLayout(JAVA_LONG.withName("rlim_cur"),
-            JAVA_LONG.withName("rlim_max"));
-
-    private static final VarHandle RLIMIT_CUR = RLIMIT_LAYOUT.varHandle(PathElement.groupElement("rlim_cur"));
-    private static final VarHandle RLIMIT_MAX = RLIMIT_LAYOUT.varHandle(PathElement.groupElement("rlim_max"));
-
     // libc is already loaded into the JVM process; defaultLookup() avoids versioning pitfalls.
     private static final SymbolLookup LIBC = LINKER.defaultLookup();
-
-    // pid_t getpid(void);
-    private static final MethodHandle getpid = LINKER.downcallHandle(LIBC.findOrThrow("getpid"),
-            FunctionDescriptor.of(JAVA_INT));
-
-    /**
-     * Calls {@code getpid()}.
-     *
-     * @return the process ID of the calling process
-     * @throws Throwable on FFM invocation error
-     */
-    public static int getpid() throws Throwable {
-        return (int) getpid.invokeExact();
-    }
 
     // thread_t thr_self(void); // Solaris-specific
     private static final MethodHandle thr_self = LINKER.downcallHandle(LIBC.findOrThrow("thr_self"),
@@ -81,58 +60,6 @@ public final class SolarisLibcFunctions extends ForeignFunctions {
      */
     public static int thr_self() throws Throwable {
         return (int) thr_self.invokeExact();
-    }
-
-    // int getrlimit(int resource, struct rlimit *rlim);
-    private static final MethodHandle getrlimit = LINKER.downcallHandle(LIBC.findOrThrow("getrlimit"),
-            FunctionDescriptor.of(JAVA_INT, JAVA_INT, ADDRESS));
-
-    /**
-     * Calls {@code getrlimit(resource, rlim)}.
-     *
-     * @param resource resource constant (e.g. {@link #RLIMIT_NOFILE})
-     * @param rlim     segment allocated with {@link #RLIMIT_LAYOUT}
-     * @return 0 on success, -1 on error
-     * @throws Throwable on FFM invocation error
-     */
-    public static int getrlimit(int resource, MemorySegment rlim) throws Throwable {
-        return (int) getrlimit.invokeExact(resource, rlim);
-    }
-
-    /**
-     * Reads {@code rlim_cur} from an rlimit segment populated by {@link #getrlimit(int, MemorySegment)}.
-     *
-     * @param rlim segment allocated with {@link #RLIMIT_LAYOUT}
-     * @return the soft resource limit
-     */
-    public static long rlimitCur(MemorySegment rlim) {
-        return (long) RLIMIT_CUR.get(rlim, 0L);
-    }
-
-    /**
-     * Reads {@code rlim_max} from an rlimit segment populated by {@link #getrlimit(int, MemorySegment)}.
-     *
-     * @param rlim segment allocated with {@link #RLIMIT_LAYOUT}
-     * @return the hard resource limit
-     */
-    public static long rlimitMax(MemorySegment rlim) {
-        return (long) RLIMIT_MAX.get(rlim, 0L);
-    }
-
-    // int gethostname(char *name, size_t namelen);
-    private static final MethodHandle gethostname = LINKER.downcallHandle(LIBC.findOrThrow("gethostname"),
-            FunctionDescriptor.of(JAVA_INT, ADDRESS, SIZE_T));
-
-    /**
-     * Calls {@code gethostname(name, namelen)}.
-     *
-     * @param name    buffer for the hostname (allocated by caller)
-     * @param namelen size of {@code name} in bytes
-     * @return 0 on success, -1 on error
-     * @throws Throwable on FFM invocation error
-     */
-    public static int gethostname(MemorySegment name, long namelen) throws Throwable {
-        return (int) gethostname.invokeExact(name, namelen);
     }
 
     // int getloadavg(double loadavg[], int nelem);
